As of now group is created and i can get group from firebase
