+++ app/src/main/res/layout/fragment_group_list.xml
@@ -17,7 +17,7 @@
             android:background="#27FFF5F5">
             <LinearLayout
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"
+                android:layout_height="0dp"
                 android:orientation="vertical"
                 android:id="@+id/linearLayoutCard"
                 app:layout_constraintTop_toTopOf="parent"
@@ -29,7 +29,7 @@
                 <androidx.recyclerview.widget.RecyclerView
                     android:id="@+id/recyclerViewsCards"
                     android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
+                    android:layout_height="match_parent"
                     android:background="@color/white"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
@@ -56,7 +56,7 @@
                 android:id="@+id/addNewGroupTextView"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:background="#B4DAD0D0"
+                android:background="#DAD0D0"
                 android:fontFamily="@font/helvetica_regular"
                 android:gravity="center"
                 android:text="Add New Group"
@@ -71,8 +71,8 @@
             <LinearLayout
                 android:layout_marginTop="5dp"
                 android:id="@+id/editTextLayout"
-                android:layout_width="407dp"
-                android:layout_height="39dp"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
 
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.0"
@@ -84,7 +84,7 @@
                 <ImageView
                     android:layout_width="30dp"
                     android:layout_height="30dp"
-                    android:layout_marginStart="25dp"
+                    android:layout_marginStart="2dp"
                     android:src="@drawable/ic_baseline_group_add_24" />
 
 
@@ -106,7 +106,9 @@
                     android:gravity="center"
                     android:hint="@string/add_group"
                     android:padding="1dp"
-                    android:textSize="20sp" />
+                    android:textSize="20sp"
+                    android:maxLines="1"
+                    android:inputType="text"/>
 
             </LinearLayout>
